[GH-73] Preparing for 3.4.0
