+++ src/commonTest/kotlin/qrcode/internals/PolynomialTest.kt
@@ -1,0 +1,6 @@
+package qrcode.internals
+
+private fun Polynomial.toList(): List<Int> = this.data.toList()
+
+private fun inputArray(vararg values: Int): IntArray =
+    IntArray(values.size) { QRMath.gexp(values[it]) }
